Archive the PHPUnit related containers. (#160)
The need for the `phpunit` containers was removed in WordPress 5.9 and only exist for running tests in old branches.

The highest version of PHP that these versions of WordPress support is PHP 8.0 required a modified version of PHPUnit 7 (see [Core-50902](https://core.trac.wordpress.org/ticket/50902)). PHPUnit 7.x is no longer supported upstream, so there will be no new releases.

With all this in mind, this removes the PHPUnit related containers from the workflow files to avoid needlessly building and publishing them. The images are being left in tact within the `images` directory in case they need to be regenerated at some point in the future.

Co-authored-by: desrosj <desrosj@git.wordpress.org>
Co-authored-by: johnbillion <johnbillion@git.wordpress.org>